refactor: Timeline 내 유틸 함수를 외부로 분리
- organizeByRounds: flat 타임라인 배열을 라운드 구조로 변환하는 로직을
  teamSelectPage/utils/organizeByRounds.ts로 분리
- formatTime: 이미 battlePage에 존재하던 공용 함수로 교체하고
  로컬 구현 및 nowRef 제거
